Доработка недочетов и ошибок.
1. Реализована основная часть, и дополнительная часть, код покрыт тестами.
2. В проекте сборщик Gradle с использованием Kotlin DSL (build.gradle.kts).
3. В репозиторий GIT запушена папка build с результатами сборки приложения. При
пересборке проекта GIT отслеживает изменения в build, отображает их на stage.
4. В классе UserServiceTest не хватало части импортов, по этой причине тесты не
собирались. После добавления импортов все тесты прошли успешно.
5. Возможны платежи на отрицательную и нулевую сумму. При отрицательной сумме
происходит пополнение баланса пользователя.
6. Нет транзакционности, блокировок и уровней изоляции. Но при этом ведется работа с
разными таблицами в одном методе, принимается решение о возможности проведения
платежа на основе текущего баланса пользователя. К чему это может привести?
7. SecurityConfig реализован корректно, особых претензий к нему нет.
8. Приложение правильно поделено на слои, но у уровня сервисов нет интерфейсов.
9. В БД на таблицах нет CHECK, нет уникальности по username. Поле gender — булево,
не совсем понятно какой пол с каким значением связан. True — male, false — female?
Единственная проверка которая есть в БД — это перечисление допустимых значений
для enum, но это произошло благодаря Hibernate DDL auto.
10. Сумма в БД представлена типом Long, и это хорошо.
11. В приложении используются DTO-классы.
12. В БД нет уникальности по логину, проверка существования пользователя с таким же
логином при его создании реализована только на стороне JAVA.
13. Использована lombok-аннотация @Data вместе с jpa-аннотацией @Entity, что вкупе с
выполненной настройкой lazy-полей @EqualsAndHashCode.Exclude не является
антипатерном.
14. Проблемы с форматирование кода. Так например нет части пробелов в конструкции ifelse (UserService.java, строки 96-112).
15. Не понятно зачем в проект добавлен файл hibernate.cfg.xml с настройками
подключения к БД, т. к. их часть дублируется в application.properties. Есть ощущение,
что все необходимые настройки можно было выполнить с помощью properties-файла,
плюс автоконфигурирование спринга.
16. Не понятно назначение роли ROLE_ADMIN, т. к. в проекте в дефолтной ветке main
эта роль нигде не прописана. В ветке master есть контроллер, который работает с ней и
дает возможность получить инфо об текущем авторизованном пользователе с этой
ролью.
Метод /api/v1/public/account (POST) - создание пользователя
 в DTO есть лишние поля (balance, payments, id), непонятно — какие обязательны для
создания пользователя, а какие — нет;
 роль пользователя передается параметром запроса, а не в его теле, хотя есть
полноценное DTO.
Из плюсов:
 есть проверка, что пользователь уже существует, но она только на стороне JAVA-кода;
 нельзя переопределить баланс значением в запросе;
 для хранения баланса используется bigint.
Метод /api/v1/user (GET) - получение баланса и логина пользователя
Из плюсов:
 берется пользователь из авторизации спринга.
Метод /api/v1/user/update (POST) - обновление данных пользователя
 нет проверки входящих данных (что email - это адрес электронной почты например),
можно сохранить пустую строку (в отличии от метода создания пользователя);
 обычно для таких действий используется http-метод PUT, а не POST.
Из плюсов:
 берется пользователь из авторизации спринга.
Метод /api/v1/user (DELETE) — удаление пользователя
Из плюсов:
 метод реализован корректно;
 берется пользователь из авторизации спринга.
Метод /api/v1/public/payment (POST) - оплата
 реализован в public-части API, возможно его использование из под анонима;
 пользователь, с баланса которого будут списаны деньги, определяется на основе поля
numberPhone, переданного в теле запроса. Т.е. аноним списывает деньги с
пользователя системы;
 нет блокировки баланса на момент проведения платежа (конкурирующее чтение);
 поле payment в таблице payment хранит в себе внешний ключ на таблицу app_user, т. е.
неочевидное название для назначения поля;
Из плюсов:
 есть проверка на то, что баланса плательщика достаточно для платежа;
 сущность платежа сохраняется через отдельный репозиторий.
Метод /api/v1/user/payments (GET) — история платежей
 значение текущей страницы передается в теле GET-запроса (да, именно так);
 нет проверки, что переданный в теле параметр больше нуля (из него вычитается
единица). Если результирующее значение page меньше нуля, то запрос завершается
ошибкой;
 нельзя управлять параметором size (количество записей на странице), он захардкожен;
 из БД за один запрос достаётся не одна страница, а вообще все записи о платежах,
которые есть в БД по этому пользователю, после чего в памяти происходит сортировка
и фильтрация полученного списка, извлечение из него подсписка отвечающего
критериям запроса (Pageable) и отдача его на фронт.
