+++ spring-basic-auth-example/src/main/java/com/example/spring_basic_auth_example/entity/Payment.java
@@ -1,5 +1,6 @@
 package com.example.spring_basic_auth_example.entity;
 
+import com.fasterxml.jackson.annotation.JsonBackReference;
 import jakarta.persistence.*;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -21,8 +22,9 @@ public class Payment {
 
     private  Long amount;
 
-    @ManyToOne(fetch = FetchType.EAGER)
-    @JoinColumn(name = "payment")
+    @JsonBackReference
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "username")
     @EqualsAndHashCode.Exclude
     @ToString.Exclude
     private User user;
