+++ spring-basic-auth-example/src/main/java/com/example/spring_basic_auth_example/model/PaymentDto.java
@@ -17,5 +17,5 @@ public class PaymentDto {
 
     private  Long amount;
 
-    private String numberPhone;
+    private String username;
 }
