+++ spring-basic-auth-example/src/main/java/com/example/spring_basic_auth_example/model/UserCreateDto.java
@@ -1,16 +1,14 @@
 package com.example.spring_basic_auth_example.model;
 
-import com.example.spring_basic_auth_example.entity.Payment;
+import com.example.spring_basic_auth_example.entity.Role;
+import com.example.spring_basic_auth_example.entity.RoleType;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
-
 import java.util.Date;
-import java.util.Set;
 
 @Data
 @RequiredArgsConstructor
 public class UserCreateDto {
-    private Long id;
 
     private String username;
 
@@ -18,18 +16,11 @@ public class UserCreateDto {
 
     private String email;
 
-    private Boolean gender;
+    private Boolean gender; //true = male, false = female
 
     private Date birthday;
 
-    private Long balance;
-
-    private Set<Payment> payments;
-
     private String password;
 
-
-
-
-
+    private RoleType roleType;
 }
