+++ spring-basic-auth-example/src/main/java/com/example/spring_basic_auth_example/security/UserDetailsServiceImpl.java
@@ -1,7 +1,7 @@
 package com.example.spring_basic_auth_example.security;
 
 
-import com.example.spring_basic_auth_example.service.UserService;
+import com.example.spring_basic_auth_example.service.UserServiceImpl;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
@@ -12,10 +12,10 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class UserDetailsServiceImpl implements UserDetailsService {
 
-    private final UserService userService;
+    private final UserServiceImpl userServiceImpl;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return new AppUserPrincipal(userService.findByUsername(username));
+        return new AppUserPrincipal(userServiceImpl.findByUsername(username));
     }
 }
