--- spring-basic-auth-example/src/main/resources/hibernate.cfg.xml
@@ -1,30 +1,0 @@
-<!DOCTYPE hibernate-configuration PUBLIC
-        "-//Hibernate/Hibernate Configuration DTD 3.0//EN"
-        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
-<hibernate-configuration>
-    <session-factory>
-        <!-- JDBC Database connection settings -->
-        <property name="connection.driver_class">org.postgresql.Driver</property>
-        <property name="connection.url">jdbc:postgresql://localhost:5432/security_app</property>
-        <property name="connection.username">postgres1</property>
-        <property name="connection.password">Bublik130698</property>
-        <!-- JDBC connection pool settings ... using built-in test pool -->
-        <property name="connection.pool_size">10</property>
-        <!-- Select our SQL dialect -->
-        <!-- Echo the SQL to stdout -->
-        <property name="show_sql">true</property>
-        <property name="hibernate.ddl-auto">update</property>
-        <!-- Set the current session context -->
-        <property name="current_session_context_class">thread</property>
-        <!-- Drop and re-create the database schema on startup -->
-        <property name="hbm2ddl.auto">validate</property>
-
-
-        <mapping class="com.example.spring_basic_auth_example.entity.User" resource="resource"/>
-        <mapping class="com.example.spring_basic_auth_example.entity.Role" resource="resource"/>
-        <mapping class="com.example.spring_basic_auth_example.entity.RoleType" resource="resource"/>
-        <mapping class="com.example.spring_basic_auth_example.entity.Payment" resource="resource"/>
-    </session-factory>
-</hibernate-configuration>
-
-<!--        jdbc:mysql://localhost:3306/skillbox?useSSL=false-->
